Refactor: class name, option, test
